backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
redis>=5.0.0
aioredis>=2.0.0  # For async Redis support
python-multipart>=0.0.5
python-docx>=0.8.11
pdfplumber>=0.10.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.21.0
scikit-learn>=1.0.0
pandas>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2
pydantic>=1.6.2,<2.0.0
pytest>=7.0.0
httpx>=0.23.0
pytest-asyncio>=0.20.0
Pillow>=10.0.0  # Required by pdfplumber
langdetect>=1.0.9
polyglot>=16.7.4
PyICU>=2.8.1  # Required for better language detection
pycld2>=0.41  # Additional language detection support
psutil>=5.9.0  # For system monitoring
prometheus-client>=0.17.0  # For Prometheus metrics format
python-magic>=0.4.27  # For file type detection
reportlab>=4.0.0  # For PDF testing
striprtf>=0.0.22  # For RTF file support
aiofiles>=23.2.1  # For async file operations
tenacity>=8.2.3  # For retries and backoff
tqdm>=4.66.1  # For progress tracking
